Conflicts:
	Makefile.am
	NEWS
	configure.ac
	src/encoder/flac_encoder.c
	src/log.c
	src/pcm_buffer.c

+++ src/stats.c
@@ -123,6 +123,6 @@ int stats_print(struct client *client)
 		      (long)g_timer_elapsed(stats.timer, NULL),
 		      (long)(pc_get_total_play_time(client->player_control) + 0.5),
 		      stats.song_duration,
-		      db_get_mtime());
+		      (long)db_get_mtime());
 	return 0;
 }
